Moved toCSV (and other relevant functions) to new file
