+++ server/routes/establishments/bulkUpload/download.js
@@ -5,6 +5,7 @@ const models = require('../../../models');
 const EstablishmentCsvValidator = require('../../../models/BulkImport/csv/establishments').Establishment;
 const WorkerCsvValidator = require('../../../models/BulkImport/csv/workers').Worker;
 const TrainingCsvValidator = require('../../../models/BulkImport/csv/training').Training;
+const WorkerCSV = require('./download/workerCSV');
 
 const { buStates } = require('./states');
 const s3 = require('./s3');
@@ -30,7 +31,7 @@ const workerCsv = async (establishments, responseSend) => {
 
   establishments.map((establishment) =>
     establishment.workers.map((worker) =>
-      responseSend(NEWLINE + WorkerCsvValidator.toCSV(establishment.LocalIdentifierValue, worker, maxQualifications)),
+      responseSend(NEWLINE + WorkerCSV.toCSV(establishment.LocalIdentifierValue, worker, maxQualifications)),
     ),
   );
 };
