+++ server/test/unit/routes/establishments/bulkUpload/download.spec.js
@@ -12,6 +12,7 @@ const mockEstablishment = require('../../../mockdata/establishment');
 const mockWorker = require('../../../mockdata/workers');
 const mockTraining = require('../../../mockdata/training');
 const WorkerCsvValidator = require('../../../../../models/BulkImport/csv/workers').Worker;
+const WorkerCSV = require('../../../../../routes/establishments/bulkUpload/download/workerCSV');
 
 describe('download', () => {
   afterEach(() => {
@@ -136,7 +137,7 @@ describe('download', () => {
     sinon.stub(WorkerCsvValidator, 'headers').callsFake((maxQuals) => {
       expect(maxQuals).to.deep.equal(9);
     });
-    sinon.stub(WorkerCsvValidator, 'toCSV').callsFake((LocalIdentifierValue, worker, maxQualifications) => {
+    sinon.stub(WorkerCSV, 'toCSV').callsFake((LocalIdentifierValue, worker, maxQualifications) => {
       expect(LocalIdentifierValue).to.deep.equal(establishment.LocalIdentifierValue);
       expect(worker).to.deep.equal(worker);
       expect(maxQualifications).to.deep.equal(9);
